proctor-service/requirements.txt
fastapi
uvicorn[standard]
opencv-python
mediapipe
ultralytics
numpy
python-dotenv
python-multipart
sqlalchemy
psycopg2-binary